collapse writer.go into wrapper.go
